// frontend/src/app/api/model/child-dto.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PersonDto } from './person-dto';


export interface ChildDto { 
    child?: PersonDto;
    guardians?: { [key: string]: Array<PersonDto>; };
}