// frontend/src/app/api/model/team-membership-dto.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PersonDto } from './person-dto';
import { TeamMemberRole } from './team-member-role';
import { TeamDto } from './team-dto';


export interface TeamMembershipDto { 
    id?: string;
    team?: TeamDto;
    person?: PersonDto;
    role?: TeamMemberRole;
    joinedAt?: string;
    leftAt?: string;
    createdAt?: string;
    updatedAt?: string;
}